OTAT-468:feature:Move ResourceNotFoundException and ResourceNotFoundAdvice to proper package
