+++ src/main/java/org/flickit/flickitassessmentcore/application/service/exception/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package org.flickit.flickitassessmentcore.application.service.assessment;
+package org.flickit.flickitassessmentcore.application.service.exception;
 
 public class ResourceNotFoundException extends RuntimeException{
     public ResourceNotFoundException(String message) {
